Support exporting installed ingredients
With this change, when cmake-cooking is invoked with the `-f` option and
the project is successfully configured, the tree of installed
ingredients will be copied (via rsync) to the specified location on the
file-system.

This feature is motivated by a work-flow for continuous
integration (CI). We may now use this option to install a project's
dependencies exactly as described in a recipe to the file-system so that
CI checks can execute without having to prepare the ingredients from
scratch every time.

This feature is not recommended for release/deployment purposes.

Fixes #2