Write `Cooking.cmake` to the build directory
Previously, when `cooking.sh` was executed it would generate
`Cooking.cmake` in the `cmake` directory at the root of the source tree.

This file never changes, and it should be treated as a build artifact in
order to be consistent with the expected behavior of out-of-source
builds.

We now write `Cooking.cmake` to the build directory, and CMake projects
which use cmake-cooking to amend the `CMAKE_MODULE_PATH` to point to
that directory.

This change also means that users of cmake-cooking do not have to have a
`cmake` directory.

Since older versions of cmake-cooking might have placed a
`Cooking.cmake` file that would conflict with the new one, we check for
the old version and delete it if necessary.

Fixes GH-3
Fixes GH-14