Assume a default recipe file
cmake-cooking now expects a recipe in a file named
`cooking_recipe.cmake` by default.

A project can still have multiple recipes, and a caller can choose with
the `-r` option as before. However, arguments to `-r` must now be a file
path instead of a name. If the path is relative, it is interpretted
relative to the source directory.

With this change, we lose the ability to run cmake-cooking without a
recipe. This seems like a reasonable trade-off.

For nested recipes, the special value of `COOKING_RECIPE <DEFAULT>`
indicates that the project should be configured with its default recipe.
Otherwise, the recipe must be specified as a file to correspond to the
changes to `-r`.

Fixes GH-13