Restore the old behavior of recipe look-up
Previously, if the argument to `-r` was "foo", then cmake-cooking would
read a recipe in a file `recipe/foo.cmake`.

With the release of 0.10.0, this behavior changed so that it was
necessary to pass the full file path to `-r` to achieve the same
behavior: `-r recipe/foo.cmake`.

This change makes sense, but the break in backwards compatibility was
poorly communicated and abrupt.

Therefore, we want to restore the previous behaviour.

This change will likely appear in the future, but hopefully more
smoothly.